feat(DataTable): fix maximes comments
