+++ packages/react/src/components/table/types.ts
@@ -1,23 +1,10 @@
 import { Cell, Column, ColumnDef, Header } from '@tanstack/react-table';
-import { CSSProperties, HTMLAttributes } from 'react';
+import { CSSProperties } from 'react';
 
 export type TableData<TData> = TData & {
     error?: boolean;
     subRows?: TableData<TData>[];
     subContent?: React.ReactNode;
-}
-
-export type TableCaptionSize = 'large' | 'medium' | 'small';
-
-export interface TableCaptionProps extends HTMLAttributes<HTMLTableCaptionElement> {
-    className?: string;
-    id?: string;
-    bold?: boolean;
-    size?: TableCaptionSize;
-}
-
-export interface StyledTableCaptionProps {
-    $bold?: boolean;
 }
 
 export type TableColumn<TData> = ColumnDef<TData> & {
